Added date/ Final commit message
Added the variable for the date. Added  The var of dateString declares the date as a text string by using the current specified locale. Added javascript to make the text string into a secondary heading format. Added the variable that Gets the day of the week for the date and for the text string to show the date of the week. Made the time of the date to change every time rendered, also to get the information inside of the function getEvent. inserted html variable and the date var before and end of the element. Added comments, ready for review.